fix: resolve all compilation errors (0.1.1)
- SurvivalSystem: MaxHunger setter for snapshot restore
- Program.cs: Fix ambiguous Version/Window with aliases, remove unused Initialize() calls
- WorldGen.cs: Fix float->int conversions for corner elevations
- NPCFlows.cs: Fix duplicate qty, unassigned var, static ErrorColor access
- Bootstrap.cs: JsonSerializer.Deserialize, DefaultBiome public setter
- Game.cs: Named tuple syntax for drawables, HandleEvent on screens, remove InputRouter SDL Event call
- TileMap.cs: RegrowTime fix, IsDepleted read-only workaround
- CHANGELOG.md: Updated for 0.1.1

+++ src/DontStarveRuneScape/Actions/ActionNotification.cs
@@ -27,7 +27,7 @@ public sealed class ActionNotification
         float duration = 2.0f)
     {
         Text = text;
-        Color = color == default ? (255, 255, 255) : color;
+        Color = color == default ? ((byte)255, (byte)255, (byte)255) : color;
         Duration = duration;
     }
